[proliferate] split connector detail modal
